docs(runtime): align capture workflow and parity docs
Why:
- restore docs and parity coverage to the runtime contracts now enforced by capture-root validation, flattened provenance artifacts, deterministic source assembly, and replay validation
- forbid older guidance that implied softer directory handling, source-level capture paths, or underspecified provenance and parity semantics
- keep the docs parity test surface within the repo's test-module size ceiling so pylint and editor diagnostics stay green

What:
- update current-state, operator guides, and mirrored workspace references to describe strict materialized capture-root inputs, capture-scoped profile inventory, flattened provenance locator columns, and rerun-safe source assembly
- align workspace inventory guidance with reducer-owned source summary and append-only capture lifecycle semantics
- split docs runtime parity coverage into a focused test package with shared support under tests/support so the assertions stay narrow and lint-clean

Checks:
- markdownlint docs/status/current-state.md docs/guides/operator-quickstart.md docs/guides/source-intake.md docs/guides/normalize-screen-stage.md docs/guides/full-operator-workflow.md docs/workspace/analysis/inventory/README.md docs/workspace/analysis/issues/README.md docs/workspace/evidence/raw/source/README.md docs/workspace/working/normalized/README.md docs/reference/wallet-inventory-artifacts.md
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run python -m tools.docs_maintenance sync --check
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run pytest -q --no-cov tests/unit/docs_runtime_parity
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run ruff check tests/support/docs_runtime_parity.py tests/unit/docs_runtime_parity
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run pylint --rcfile=.pylintrc-tests tests/support/docs_runtime_parity.py tests/unit/docs_runtime_parity/test_routes.py tests/unit/docs_runtime_parity/test_runtime_docs.py tests/unit/docs_runtime_parity/test_repo_doc_guards.py
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run python -m tools.run_quality_gates --full-tests
- UV_PROJECT_ENVIRONMENT=$HOME/.venvs/tallylot-py312 uv run python -m tools.run_ci_parity_checks
