+++ docs/guides/full-operator-workflow.md
@@ -46,6 +46,9 @@ part of the current round-close process.
 - Keep untouched statements, HTML exports, and required upstream sidecars in
   raw evidence.
 - Build the capture manifest and profile the settled capture.
+- `source profile` and `source normalize` must point at the materialized
+  capture root with matching `capture.json` metadata. They do not accept source
+  roots or arbitrary directories.
 - Use [Source Intake](source-intake.md) for the full command sequence and
   artifact review points.
 
@@ -53,8 +56,13 @@ part of the current round-close process.
 
 - Normalize the settled capture and review the emitted issues and summaries
   under `working/normalized/captures/<capture_uid>/`.
+- Review `profile_inventory.csv` as the capture-scoped discovery contract for
+  statement-backed evidence, archive-member provenance, and issue-context
+  resolution.
 - Run `source assemble` before reconciliation so the accepted capture outputs
   land under `working/normalized/sources/<source>/`.
+- `source assemble` owns that source dataset surface and rewrites its known
+  generated files on rerun instead of leaving stale assembled artifacts behind.
 - Use normalization-owned `balances.csv` and `balance_evidence.csv` directly
   when the source adapter already produced canonical balance artifacts.
 - When balances need to be authored manually, run
