+++ docs/workspace/analysis/inventory/README.md
@@ -19,8 +19,22 @@ Files:
 - `location_inventory_summary.json`
 
 `source_captures.csv` is the append-only capture registry keyed by
-`capture_uid`. `location_inventory.csv` and the related evidence and issue
-artifacts remain derived outputs rebuilt from normalized datasets.
+`capture_uid`. Each row records one lifecycle event for one capture, including
+duplicate-blocked or excluded outcomes that never materialize a new raw capture
+root. `location_inventory.csv` and the related evidence and issue artifacts
+remain derived outputs rebuilt from normalized datasets.
+
+`location_inventory_evidence.csv` carries the shared flattened provenance
+locator family with `evidence_` prefixes:
+
+- `evidence_capture_uid`
+- `evidence_relative_path`
+- `evidence_archive_member_path`
+- `evidence_locator_kind`
+- `evidence_anchor`
+
+Those columns identify the supporting capture-scoped evidence row or document.
+They replace ad hoc source-level path assumptions.
 
 Refresh with:
